Plot mean-state field map
This script reads reanalyses and CMIP6 datasets as an ensemble,
and calculates their ensemble mean.
It also addresses the mean PPT field south of 45• and the difference
of each dataset from the ens. mean.
In this script, the ens. mean map is overlaped with the models agreement
as stippling - model agreement is defined by the ensemble size through
binomial distribution. For achieving 95% confidence level on agreement, 6/7
reanalyses and 15/23 models must agree.
At last, this script plots: (i) map of all-time average ppt for each dataset
                            (ii) map of the 50yr difference for each dset